training/requirements.txt
albumentations==1.4.1
python-dotenv==1.1.1
pytorch_lightning
pynrrd
accelerate==1.10.1
datasets==4.1.1
huggingface-hub==0.35.2
loguru==0.7.3
monai==1.5.1
nibabel==5.3.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
pandas==2.3.2
pillow==11.3.0
pyarrow==21.0.0
pydicom==3.0.1
safetensors==0.6.2
simpleitk==2.5.2
tokenizers==0.22.1
torch==2.6.0
torchio==0.20.22
torchvision==0.21.0
matplotlib
# Editable install with no version control (transformer-maskgit==0.0.0)
-e /training/CT-CLIP/transformer_maskgit
transformers==4.56.2
vector-quantize-pytorch==1.1.2
